Fixed warnings from -Wall on 64-bit
I hadn't been compiling with -Wall. Three remain in
ExecPE.C for work in progress.


git-svn-id: file:///usr/casc/overture/ROSE/svn/ROSE/trunk/ROSE@696 a27c51de-2d2b-4d1b-a6e8-f7f87814f453

+++ src/frontend/ExecFormats/ExecGeneric.C
@@ -224,7 +224,7 @@ ExecFile::find_holes()
     std::vector<std::pair<addr_t,addr_t> > extents;
 
     addr_t offset = 0;
-    while (offset < sb.st_size) {
+    while (offset < (uint64_t)sb.st_size) {
         std::vector<ExecSection*> sections = lookup_section_offset(offset, 1);
         if (0==sections.size()) {
             addr_t next_offset = lookup_next_offset(offset);
